ADD alias for removing conflicting labels.
After configuring a Waffle.io board with labels "Status - workable" and "Status - up next", respectively on the boards Ready" or "Up Next" board. Waffle.io will set these labels to an issue, when it's manually moved to either of these boards.

When the "git work-on 123" command is invoked, ghi will assign the "Status - work in progress" to the issue, but the issue won't move on the board, because it's already in a column corresponding to a label on it, i.e. ".. - workable" or ".. - up next".

This alias will remove the "workable" and "up next" from the issue, irrespectively of whether they're set on it.

It also makes sense to remove this from an agile point of view,
- it it's being worked on, it's definitely ready, and 
- it's also not "up next" anymore; it's in progress.